Move chain types from `deposit.ts` to `lib/contracts`
Those types are purely chain-related and should be placed next to
the respective contracts interfaces.
